# Conflicts:
#	client/src/main/java/ar/edu/pod/tp2/client/Queries/Query1.java
#	scripts/query1.sh

+++ api/src/main/java/ar/edu/pod/tp2/Reducer/Query1Reducer.java
@@ -3,14 +3,14 @@ package ar.edu.pod.tp2.Reducer;
 import com.hazelcast.mapreduce.Reducer;
 import com.hazelcast.mapreduce.ReducerFactory;
 
-public class Query1Reducer implements ReducerFactory<String, Integer, Long> {
+public class Query1Reducer implements ReducerFactory<String, Long, Long> {
 
     @Override
-    public Reducer<Integer, Long> newReducer(String s) {
+    public Reducer<Long, Long> newReducer(String s) {
         return new query1Class();
     }
 
-    private static class query1Class extends Reducer<Integer,Long>{
+    private static class query1Class extends Reducer<Long,Long>{
         private Long sum;
 
         @Override
@@ -19,7 +19,7 @@ public class Query1Reducer implements ReducerFactory<String, Integer, Long> {
         }
 
         @Override
-        public void reduce(Integer integer) {
+        public void reduce(Long integer) {
             sum += integer;
         }
